Rethink ring handling and call status monitor
Since it is expected for a daemon to be detecting rings, it raises the
question of how a user would be graphically notified, especially in a
multiuser environment. So now the ring process has been lifted to a
mqbus-based ringd which will handle notifying users and immediate
actions.

Also list-calls was added to list the status of current calls which
includes incoming calls that have not yet been accepted. Scripts can
query this to get all call status and to also filter for a specific
status