Allow arbitrary comment before every generated sms AT command
When communicating to ttyio.c, it is possible to pass a label/comment
that allows the caller to know the result of their cmd. Previously this
comment was just prepended before sms was run. However, this didn't work
well with concatenated SMS messages as multiple AT commands would be
generated. Now the sms takes an option to just append an arbitrary
string before every AT command it generates.